+++ resources/js/admin/view/main/reviews/ReviewsReported.vue
@@ -165,12 +165,14 @@
         let indexItemsData = this.itemsData.findIndex(a => a.id == e.id)
         axios.patch(`/api/review/${e.id}`, {
             review_status_id: e.review_status_id,
+            rating: e.rating,
             title: e.title,
             description: e.description,
           })
           .then((response) => {
             this.itemsData[indexItemsData].review_status = new_review_status
             this.itemsData[indexItemsData].review_status_id = newData.value
+            this.itemsData[indexItemsData].rating = e.rating
             this.itemsData[indexItemsData].title = e.title
             this.itemsData[indexItemsData].description = e.description
             this.$refs.reviewsModal.modalShow(false)