+++ resources/views/pages/user-write-review.blade.php
@@ -15,14 +15,19 @@
           @endslot
           @slot('image_area')
           <div class="bg-white  rounded" style="height: auto;width:120px">
-            <img class="img-fluid " src="{{ asset('images/websites/logo-company.png') }}" alt="">
+            @if ($company->avatar)
+
+            <img class="img-fluid " src="{{ asset("/storage/company/{$company->avatar}") }}" alt="">
+            @else
+            <span class="b-avatar badge-secondary rounded-sm" style="width:120px;height:120px"><i class="fa fa-briefcase fa-3x"></i></span>
+            @endif
           </div>
           @endslot
           @slot('data_area')
           <div>
             <div class="text-light text-center text-sm-left">{{$company->business_category->name}}</div>
             <div class="text-white text-center text-sm-left">
-              <h3>{{$company->name}}</h3>
+              <h1>{{$company->name}}</h1>
             </div>
             <div class="d-flex mb-3 justify-content-center justify-content-sm-start">
               @php
@@ -36,6 +41,10 @@
                 @for ($a = 1 ; $a <= (5 - $rating); $a++ ) <x-rating-grey>
                   </x-rating-grey>
                   @endfor
+                     <div class="text-light">
+
+                       Rating : {{$company->rating}} of 5.0
+                     </div>
             </div>
             <div class="d-flex">
               <div class="text-success">Solutions : {{$company->review_solution ?? 0}}</div>
@@ -49,10 +58,10 @@
 
       </div>
       <div class="col-lg-4 ">
-        <div>
-                    @php
-                    $total_star_count = $company->star_1 + $company->star_2 + $company->star_3 + $company->star_4 + $company->star_5;
-                    @endphp
+        <div class="mb-3">
+          @php
+          $total_star_count = $company->star_1 + $company->star_2 + $company->star_3 + $company->star_4 + $company->star_5;
+          @endphp
           <x-progress>
             @slot('value')
             {{ $total_star_count ? ($company->star_5 * 100) /($total_star_count) : 0 }}
@@ -84,6 +93,9 @@
             1 stars
           </x-progress>
         </div>
+           <div class="w-100 text-center text-light">
+             Visitor : {{$company->visited}}
+           </div>
       </div>
     </div>
   </x-header-wrapper>
@@ -93,9 +105,9 @@
       <div class="row">
         <div class="col-lg-8">
           @if ($reviewed)
-<div class="alert alert-info">
-Company Reviewed
-</div>
+          <div class="alert alert-info">
+            Company Reviewed
+          </div>
           @else
           <user-write-review :companyid='@json($company->id)' :slug='@json($company->slug)' :companyname='@json($company->name)' :reviewtype='@json($review_type)' :term='@json($term)'>
             @endif